+++ lib/akasha/changeset.rb
@@ -2,17 +2,22 @@ module Akasha
   # Represents changes to an aggregate, for example an array of
   # events generated when handling a command.
   class Changeset
-    attr_reader :aggregate_id, :events
+    attr_reader :events
 
-    def initialize(aggregate_id)
-      @aggregate_id = aggregate_id
+    def initialize(aggregate)
+      @aggregate = aggregate
       @events = []
+    end
+
+    def aggregate_id
+      @aggregate.id
     end
 
     # Adds an event to the changeset.
     def append(event_name, **data)
       id = SecureRandom.uuid
-      event = Akasha::Event.new(event_name, id, { aggregate_id: @aggregate_id }, **data)
+      event = Akasha::Event.new(event_name, id, { aggregate_id: aggregate_id }, **data)
+      @aggregate.apply_events([event])
       @events << event
     end